fix: update types for APOD and EPIC and LAUNCH due to type changes on the external APIs

+++ src/epic/entities/epic.entity.ts
@@ -1,11 +1,11 @@
-import { ObjectType, Field, Int } from '@nestjs/graphql';
+import { ObjectType, Field, Float } from '@nestjs/graphql';
 import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity()
 @ObjectType()
 export class Epic {
   @PrimaryGeneratedColumn()
-  @Field(() => Int, { description: 'Example field (placeholder)' })
+  @Field(() => Float, { description: 'Example field (placeholder)' })
   identifier: number;
 
   @Column()
